add update event functionalities
Teacher clicks on  the update button of the event to be updated. Then a form which is prefilled with the current data of that event appears. Teacher makes the changes and clicks on the update button. When the teacher clicks on the return to events button, the updated version of the event list is shown.