update models and add create operations for training-skill- calendar
user, training, skill and studentcalendar models are updated. Teacher can create trainings via a form and choose students to add those trainings to the selected students. When the teacher selects students, for the selected student, some skill objects with the date and the string in the skills field of that training. Besides skill objects, a studentcalendar object is created with the date of the training and list of skill objects.